+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Remove `unrecognized` status from Transaction, Subscription, and CreditCardVerification
 * Remove `GrantedPaymentInstrumentUpdate` kind from Webhook
 * Remove Coinbase references
+* Add GatewayTimeout exception
+* Rename DownForMaintenance exception to ServiceUnavailable
 
 ## 3.59.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors